Move tests to tested file's location
