+++ src/client/App.test.tsx
@@ -1,9 +1,16 @@
 import React from "react";
 import { render, unmountComponentAtNode } from "react-dom";
-import { About } from "./../../containers";
+import { BrowserRouter } from "react-router-dom";
+
+import App from "./App";
 
 it("renders without crashing", () => {
   const div = document.createElement("div");
-  render(<About />, div);
+  render(
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>,
+    div
+  );
   unmountComponentAtNode(div);
 });
